Refactored components folder and added form validity to all of the menu forms
