+++ src/App.css
@@ -5,6 +5,7 @@
 --midgrey: #D9D9D9;
 --darkgrey: #393939;
 --violet: #8a4fec;
+--green: #45EA56 ;
 }
 
 html {
