+++ src/pages/Profile.jsx
@@ -2,16 +2,16 @@ import Container from "react-bootstrap/Container";
 import Row from "react-bootstrap/Row";
 import Col from "react-bootstrap/Col";
 import { useUser } from "../contexts/UserProvider";
-import ConversationList from "../components/ConversationList";
+import ConversationList from "../components/MenuColumn/Conversations/ConversationList";
 import OpenConversation from "../components/ChatColumn/OpenConversation";
 import { useState } from "react";
-import Contacts from "../components/Contacts";
+import Contacts from "../components/MenuColumn/Contacts/Contacts";
 import Timer from "../components/RightColumn/Timer";
 import ConvoInfo from "../components/RightColumn/ConvoInfo";
-import Settings from "../components/Settings";
+import Settings from "../components/MenuColumn/Settings/Settings";
 import styled from "styled-components";
-import UserMenu from "../components/UserMenu";
-import { ConversationUserList } from "../components/ConversationUserList";
+import UserMenu from "../components/MenuColumn/UserMenu";
+import { ConversationUserList } from "../components/ContactsColumn/ConversationUserList";
 
 const Profile = () => {
     const [menu, setMenu] = useState("Conversations");
